feat(aot): Soft epoch-invariant consecutive-dirty fuse + joint epoch bump static gate (#2693, refine #2640 #2668 #2541 #2046)
P1 [production-readiness] Soft epoch-invariant observability +
joint-epoch-bump static gate.

A) Consecutive-dirty Soft fuse (#2693 §A):
  - After each Soft walk (periodic + event-driven), track
    g_consecutive_dirty_count via aura_2693_soft_fuse_record.
  - behind_after_clear > 0 → fetch_add(1); == 0 → store(0).
  - When consec >= K (default K=3, env AURA_EPOCH_INVARIANT_SOFT_FUSE_K;
    K=0 disables) → bump g_2693_soft_fuse_fallback_total +
    epoch_invariant_soft_fuse_total.
  - Soft zero-cost on quiet path: single atomic op per walk;
    K=0 disables the fuse entirely.
  - Optional one-shot force-JIT region seed deferred to follow-up.

B) Joint epoch bump static gate (#2693 §B):
  - scripts/coverage/checks/check_joint_epoch_bump_coverage.py forbids
    bare g_current_bridge_epoch.fetch_add / .store and
    g_aot_table_epoch.fetch_add / .store outside the documented
    lockstep helpers (atomic_bump_epochs_and_stamp_bridge in
    service.ixx + aura_aot_bump_func_table_epoch +
    commit_func_table_swap + aura_set_current_bridge_epoch) and the
    bridge TU / stub / header / aot_mangle.h allow-list.
  - --self-test mode demonstrates bad patch caught + allow-list passes.
  - Wired into build.py gate next to check_2668_coverage.

Query surface (additive, no #2640/#2668/#2366/#2541/#2304 regression):
  epoch-invariant-soft-fuse-total,
  epoch-invariant-consecutive-dirty-total,
  epoch-invariant-soft-fuse-k-default,
  epoch-invariant-soft-fuse-wired,
  schema-2693, issue-2693.

Tests extended per #81967 — tests/compiler/test_epoch_invariant_walk.cpp
gains AC1-6 source-cite + linter --self-test block (no new file,
no docs/design/2693-* per #1655).

Build hygiene: capability_model.hh kCapabilityProductionDefaultIssue
gains 'static' (was inline constexpr as non-static data member,
illegal C++26 — broken at origin/main before #2693). Pre-existing
build errors in evaluator_mutation_boundary.cpp (DrainReason,
compact_r) + evaluator_primitives_obs_jit.cpp (g_shape_storm_*
namespace, maybe_stamp_stable_ref_isolation_tenant) remain as
follow-up — not introduced by #2693.

+++ src/compiler/evaluator_mutation_boundary.cpp
@@ -961,8 +961,8 @@ Evaluator::MutationCheckpoint Evaluator::exit_mutation_boundary(bool success) {
     // `double_drain_prevented` to surface the race. Closes the residual
     // unhealed window from novel interleavings (#2690).
     if (!nested_boundary && success) {
-        aura_hot_update_drain_pending_recovery(static_cast<std::uint8_t>(
-            DrainReason::BoundaryExit));
+        aura_hot_update_drain_pending_recovery(
+            static_cast<std::uint8_t>(DrainReason::BoundaryExit));
     }
     // Issue #2604: outermost MutationBoundary exit auto-drain deferred
     // reemit + one region-filtered pass. Closes the "visible but
@@ -2473,10 +2473,10 @@ Evaluator::MutationBoundaryGuard::~MutationBoundaryGuard() {
                                                        // already covers the
                                                        // user-visible gate
                 pin_contract_held,
-                /*root_remap_stable_ref_fail_total=*/static_cast<std::uint64_t>(
-                    compact_r.root_remap_stable_ref_fail_total),
-                /*root_remap_closure_capture_fail_total=*/static_cast<std::uint64_t>(
-                    compact_r.root_remap_closure_capture_fail_total),
+                /*root_remap_stable_ref_fail_total=*/
+                static_cast<std::uint64_t>(compact_r.root_remap_stable_ref_fail_total),
+                /*root_remap_closure_capture_fail_total=*/
+                static_cast<std::uint64_t>(compact_r.root_remap_closure_capture_fail_total),
                 /*objects_moved=*/static_cast<std::uint64_t>(densify_objects_moved),
                 /*untracked_kept_count=*/static_cast<std::uint64_t>(densify_untracked_kept));
         if (moving_unified_success) {
